### 2019-10-05 #1786 @kyleniemeyer
Hi @pssachdeva, thanks for your interest in JOSS! It looks like the article has one or two problems preventing compilation:
- in the YAML header, please add a date field: `date: 5 August 2019` (e.g.)
- author entries with multiple affiliations need those wrapped in quotation marks, so please wrap the affiliations for authors 1, 2, and 4 (e.g., `affiliation: 1, 2, 3` -> `affiliation: "1, 2, 3"`)
