### 2019-10-05 #1786 @kyleniemeyer
@pssachdeva ah, unfortunately I don't believe our template supports the `header-includes:` command.

@arfon, can you confirm? It looks like they want to use an `algorithm` LaTeX environment:
```latex
\begin{algorithm}[t]
    \caption{\textsc{UoI-Lasso}}
    \label{alg:uoi}
    \hspace*{\algorithmicindent} \textbf{Input}:
    $X \in \mathbb{R}^{N\times p}$ design matrix \\
    \hspace*{4.5em} $y \in \mathbb{R}^{N}$ response variable \\
    \hspace*{4.5em} Regularization strengths $\left\{\lambda_j \right\}_{j=1}^{q}$ \\
    \hspace*{4.5em} Number of resamples $N_S$ and $N_E$ \\
    \hspace*{4.5em} Loss function $L(\beta; X, y)$

    \begin{algorithmic}[1]
        \Statex \textit{Model Selection}
        \For{$k = 1$ to $N_S$}
            \State Generate resample $X^k$, $y^k$
            \For{$j=1$ to $q$}
                \State $\hat{\beta}^{jk}\leftarrow$ Lasso regression (penalty $\lambda_j$) of $y^k$ on $X^k$
                \State $S_j^k \leftarrow \left\{i\right\}$ where $\hat{\beta}^{jk}_i \neq 0$
            \EndFor
        \EndFor
        \For{$j = 1$ to $q$}
            \State $\displaystyle S_j \leftarrow \bigcap_{k=1}^{N_S}S_j^k$ \hfill $\triangleright$ \textit{ Intersection}
        \EndFor
        \State \textit{Model Estimation}
        \For{$k=1$ to $N_E$}
            \State Generate training $\left(X_T^k, y_T^k\right)$ and evaluation $\left(X_E^k, y_E^k\right)$ resamples
            \For{$j=1$ to $q$}
                \State $X_{T, j}^k, X_{E, j}^k \leftarrow $ $X_T^k, X_E^k$ with features $S_j$ extracted.
                \State $\hat{\beta}^{jk} \leftarrow$ OLS Regression of $y_T^k$ on $X_{T,j}^k$
                \State $\ell^{jk} \leftarrow L(\hat{\beta}^{jk}; X^k_{E, j}, y_E^k)$
            \EndFor
            \State $\hat{\beta}^k \leftarrow \underset{\hat{\beta}^{jk}}{\text{argmin}} \ \ell^{jk}$
        \EndFor
        \State $\hat{\beta}^* = \underset{k}{\text{median}}\left(\hat{\beta}^k\right)$ \hfill $\triangleright$ \textit{ Union} \\
        \Return $\hat{\beta}^*$
    \end{algorithmic}
\end{algorithm}
```
