### 2019-10-05 #1786 @pssachdeva
@kyleniemeyer Hm, OK. If we definitely can't use the `algorithm` package, perhaps we could include a picture of the pseudocode instead?
