### 2019-10-05 #1786 @arfon
@pssachdeva is it just the `algorithms` package you need here?
