### 2019-10-06 #1786 @pssachdeva
Right now it's using `algorithms` and `algpseudocode`, but we could switch to something else if that would make it easier.
