+++ .travis.yml
@@ -1,11 +1,19 @@
 language: python
+os:
+    - linux
+    - osx
 python:
     - 3.5
     - 3.6
     - 3.7
+compiler:
+    - clang
+    - gcc
 before_install:
-    - sudo apt-get update
-    - sudo apt-get install -y openmpi-bin libopenmpi-dev gcc
+    - if [ "$TRAVIS_OS_NAME" = "osx" ]; then brew update          ; fi
+    - if [ "$TRAVIS_OS_NAME" = "osx" ]; then brew install openmpi; fi
+    - if [ "$TRAVIS_OS_NAME" = "linux" ]; then sudo apt-get update; fi
+    - if [ "$TRAVIS_OS_NAME" = "linux" ]; then sudo apt-get install -y openmpi-bin libopenmpi-dev gcc; fi
 install:
     - pip install -r requirements-dev.txt
     - pip install codecov
