+++ .travis.yml
@@ -1,19 +1,31 @@
 language: python
-os:
-    - linux
-    - osx
+os: linux
 python:
-    - 3.5
-    - 3.6
-    - 3.7
-compiler:
-    - clang
-    - gcc
+  - 3.5
+  - 3.6
+  - 3.7
+matrix:
+  include:
+    - os: osx
+      language: generic
+      env: PYTHON=3.7.4
 before_install:
-    - if [ "$TRAVIS_OS_NAME" = "osx" ]; then brew update          ; fi
-    - if [ "$TRAVIS_OS_NAME" = "osx" ]; then brew install openmpi; fi
-    - if [ "$TRAVIS_OS_NAME" = "linux" ]; then sudo apt-get update; fi
-    - if [ "$TRAVIS_OS_NAME" = "linux" ]; then sudo apt-get install -y openmpi-bin libopenmpi-dev gcc; fi
+    - |
+      if [ "$TRAVIS_OS_NAME" = "osx" ]; then
+        brew update
+        brew install openmpi pyenv-virtualenv
+        pyenv install --list
+        pyenv install $PYTHON
+        export PYENV_VERSION=$PYTHON
+        export PATH="/Users/travis/.pyenv/shims:${PATH}"
+        pyenv virtualenv venv
+        source /Users/travis/.pyenv/versions/3.7.4/envs/venv/bin/activate
+      fi
+    - |
+      if [ "$TRAVIS_OS_NAME" = "linux" ]; then
+        sudo apt-get update
+        sudo apt-get install -y openmpi-bin libopenmpi-dev gcc
+      fi
 install:
     - pip install -r requirements-dev.txt
     - pip install codecov
