### 2019-10-06 c21ca89
Update installation.rst
