### 2019-10-07 3bd2c78
Update MANIFEST.in
