### 2019-10-07 001c129
Update MANIFEST.in
