### 2019-10-07 17326f3
Update MANIFEST.in
