+++ MANIFEST.in
@@ -2,3 +2,6 @@ include pyproject.toml
 include requirements.txt
 include requirements-dev.txt
 include LICENSE.txt
+include liblbfgs/*.h
+include liblbfgs/*.c
+include pyuoi/lbfgs/*pyx
