### 2019-10-07 0959c2c
Update MANIFEST.in
