### 2019-10-07 2cac3f7
Update MANIFEST.in
