### 2019-10-07 5dd3f25
Update MANIFEST.in
