### 2019-10-07 7a94d84
made pycasso optional
