+++ tests/test_uoi_lasso.py
@@ -1,3 +1,4 @@
+import pytest
 import numpy as np
 from numpy.testing import (assert_array_equal, assert_array_almost_equal_nulp,
                            assert_equal, assert_allclose)
@@ -6,6 +7,10 @@ from sklearn.datasets import make_regression
 from sklearn.linear_model import Lasso
 from sklearn.metrics import r2_score
 from sklearn.linear_model.coordinate_descent import _alpha_grid
+try:
+    import pycasso
+except ImportError:
+    pycasso = None
 
 from pyuoi import UoI_Lasso
 from pyuoi.linear_model.lasso import PycLasso
@@ -92,15 +97,16 @@ def test_uoi_lasso_toy():
 
     assert_allclose(lasso.coef_, beta)
 
-    lasso = UoI_Lasso(
-        fit_intercept=False,
-        selection_frac=0.75,
-        estimation_frac=0.75,
-        solver='pyc'
-    )
-    lasso.fit(X, y)
+    if pycasso is not None:
+        lasso = UoI_Lasso(
+            fit_intercept=False,
+            selection_frac=0.75,
+            estimation_frac=0.75,
+            solver='pyc'
+        )
+        lasso.fit(X, y)
 
-    assert_allclose(lasso.coef_, beta)
+        assert_allclose(lasso.coef_, beta)
 
 
 def test_get_reg_params():
@@ -201,6 +207,7 @@ def test_fit_intercept():
     assert not lasso._estimation_lm.fit_intercept
 
 
+@pytest.mark.skipif(pycasso is None, reason='pycasso not installed')
 def test_choice_of_solver():
     '''Tests whether one can correctly switch between solvers in UoI Lasso'''
 
@@ -211,6 +218,17 @@ def test_choice_of_solver():
     assert(isinstance(uoi2._selection_lm, PycLasso))
 
 
+@pytest.mark.skipif(pycasso is not None, reason='pycasso is installed')
+@pytest.mark.xfail(raises=ModuleNotFoundError)
+def test_pycasso_error():
+    """Tests whether an error is raised if pycasso is not installed.
+    """
+
+    uoi2 = UoI_Lasso(solver='pyc')
+    assert(isinstance(uoi2._selection_lm, PycLasso))
+
+
+@pytest.mark.skipif(pycasso is None, reason='pycasso not installed')
 def test_pyclasso():
     """Tests whether the PycLasso class is working"""
 
