+++ .travis.yml
@@ -29,6 +29,7 @@ install:
     - pip install -r requirements-dev.txt
     - pip install codecov
     - pip install mpi4py
+    - pip install pycasso
     - python setup.py build
     - python setup.py develop
     - pip install pytest-cov
