### 2019-10-07 6dc4269
added a few tests
