### 2019-10-08 f57c0f4
add logger test
