### 2019-10-09 572c48a
fix intercept calculation in post_fit and handled edge cases with constant or zero response variable
