+++ pyuoi/linear_model/base.py
@@ -205,8 +205,13 @@ class AbstractUoILinearModel(SparseCoefMixin, metaclass=_abc.ABCMeta):
 
         # extract model dimensions
         n_features = X.shape[1]
-
         n_coef = self.get_n_coef(X, y)
+
+        if np.unique(y).size == 1:
+            self.coef_ = np.zeros((self.output_dim, n_features))
+            self._fit_intercept(X, y)
+            self._post_fit(X, y)
+            return self
 
         ####################
         # Selection Module #
