### 2019-10-09 ec32c1a
Update pyuoi/linear_model/base.py
Co-Authored-By: Jesse Livezey <jesse.livezey@gmail.com>
