### 2019-10-10 #1799 @whedon
[REVIEW]: PyUoI: The Union of Intersections Framework in Python
**Submitting author:** @pssachdeva (<a href="http://orcid.org/0000-0002-6809-2437">Pratik Sachdeva</a>)
**Repository:** <a href="https://github.com/BouchardLab/pyuoi" target ="_blank">https://github.com/BouchardLab/pyuoi</a>
**Version:** 1.0.0
