### 2019-10-10 345d92a
Update installation.rst
