### 2019-11-10 #1799 @yzhao062
@pssachdeva (@terrytangyuan for some reason I could mention Pratik in this thread).

It is a good package and I just did some initial tests and crosses some items from the list. Currently, I have three major questions:

1. License: not necessarily an OSI approved one but the term is consistent with the open-source standard. Or could you clarify if it is a modified version of one of the OSI licenses?

2. I could run examples successfully (plot_uoi_lasso.py), but I am not sure why the figure is not showing up (I will try to figure it out on my end). Another note is it will be helpful to have an example folder on GitHub so please could find it easily. It took me some time to navigate.

3. I tried to run the unit test and doc test, but failed due to the import errors. I would not worry too much about the unit test as CI takes care of that. However, could you provide the result for doc test. You could run "pytest --doctest-modules pyuoi/" and address any potential errors there (if there is any).

Other than that, the installation has instruction and it is easy (via pip). CI is enabled and codecov is in place. I could also find the contribution guide (https://pyuoi.readthedocs.io/en/latest/contributing.html).

I look forward to your confirmation on the license, example place/instruction, and the result of the doc test.

I will go through the rest review processes at the same time.

Cheers!
