### 2019-11-11 1e24dc6
Fixed doc errors and added docs tests
