+++ docs/source/pyuoi/datasets/datasets.rst
@@ -1,0 +1,13 @@
+===========
+pyuoi.datasets
+===========
+
+Dataset utility functions for the ``pyuoi`` package.
+
+Testing Utilities
+-----------------
+
+.. automodule:: pyuoi.datasets
+    :noindex:
+    :members: make_linear_regression, make_classification,
+              make_poisson_regression
