+++ .travis.yml
@@ -36,5 +36,7 @@ install:
 script:
     - flake8 pyuoi tests docs/gallery
     - pytest --cov=./ tests
+    - pytest --doctest-modules pyuoi
+    - sphinx-build -nW -b html docs/source /docs/build
 after_success:
     - codecov
