+++ .travis.yml
@@ -37,6 +37,6 @@ script:
     - flake8 pyuoi tests docs/gallery
     - pytest --cov=./ tests
     - pytest --doctest-modules pyuoi
-    - sphinx-build -nW -b html docs/source /docs/build
+    - sphinx-build -nW -b html docs/source docs/build
 after_success:
     - codecov
