+++ docs/source/index.rst
@@ -27,7 +27,7 @@ Further details on the UoI framework can be found in [Bouchard2017]_ and
 
    introduction
    installation
-   examples/index
+   auto_examples/index
    contributing
    mpi
    api
