+++ docs/source/pyuoi/utils.rst
@@ -10,14 +10,6 @@ Scoring Utilities
 .. automodule:: pyuoi.utils
    :members: AIC, BIC, AICc, log_likelihood_glm
 
-Testing Utilities
------------------
-
-.. automodule:: pyuoi.utils
-    :noindex:
-    :members: make_linear_regression, make_classification,
-              make_poisson_regression
-
 Other Utilities
 ---------------
 
