### 2019-11-11 441e32d
added plt.show
