### 2019-11-11 #1799 @pssachdeva
Hi @yzhao062,

Thanks for your comments.

> License: not necessarily an OSI approved one but the term is consistent with the open-source standard. Or could you clarify if it is a modified version of one of the OSI licenses?

We used the _Lawrence Berkeley National Labs BSD variant license_, whose short identifier is "BSD-3-Clause-LBNL". The license is the same as BSD-3-Clause, but with an additional default contribution licensing clause. See [here](https://spdx.org/licenses/BSD-3-Clause-LBNL.html) for more details.

> I could run examples successfully (plot_uoi_lasso.py), but I am not sure why the figure is not showing up (I will try to figure it out on my end).

This might be happening because we didn't have a `plt.show()` at the end of the file. Our latest pull request added that to the end of the example scripts, so hopefully it should work now.

> Another note is it will be helpful to have an example folder on GitHub so please could find it easily. It took me some time to navigate.

Thanks for pointing this out. We reorganized our Sphinx Gallery directories so that the scripts are located in a top-level `examples` directory.

> I tried to run the unit test and doc test, but failed due to the import errors.

Could you tell us how you installed the package and which import errors you received? Depending on the install method, we might need to update our `requirements.txt` or `requirements-dev.txt` to make sure all the necessary packages are installed.

> However, could you provide the result for doc test. You could run "pytest --doctest-modules pyuoi/" and address any potential errors there (if there is any).

We don't have any code in our docstrings that needs to be tested, so running that command returns no errors with no tests run. However, this point did make us realize that our continuous integration had not been testing whether the documentation was being compiled correctly. Thus, our newest pull request should add that functionality to our pipeline.
