### 2019-11-14 7e5fc5b
only allow effectivly 1d inputs for y in linear regression
