### 2019-11-14 c6c35d6
added one more y shape check
