### 2019-11-14 3deec36
added more shape checks to _score_predictions
