+++ pyuoi/linear_model/base.py
@@ -603,6 +603,16 @@ class AbstractUoILinearRegressor(AbstractUoILinearModel,
         # Select the data relevant for the estimation_score
         X = X[boot_idxs[self._estimation_target]]
         y = y[boot_idxs[self._estimation_target]]
+
+        if y.ndim == 2:
+            if y.shape[1] > 1:
+                raise ValueError('y should either have shape ' +
+                                 '(n_samples, ) or (n_samples, 1).')
+            y = np.squeeze(y)
+        elif y.ndim > 2:
+            raise ValueError('y should either have shape ' +
+                             '(n_samples, ) or (n_samples, 1).')
+
         y_pred = fitter.predict(X[:, support])
         if y.shape != y_pred.shape:
             raise ValueError('Targets and predictions are not the same shape.')
