### 2019-11-22 #1799 @puolival
**Checklist**

The software is well written and documented. I did encounter some issues, but I think they should be easy to solve, and seem to be related to first uses outside the original development environment. In particular, the installation failed for me, but I was still able to work around it by manual work and testing several different environments. I am quite sure that all issues I faced were due to conflicting version numbers. Please make it visible on the requirement lists and the installation instructions page which version numbers of each used package are supported, and which Python version the user is supposed to have.

To give an example, in one of my environments, I received errors like this:
`In [1]: import pyuoi
  File "/usr/local/.../pyuoi/linear_model/base.py", line 19
    class AbstractUoILinearModel(SparseCoefMixin, metaclass=_abc.ABCMeta):
                                                           ^
SyntaxError: invalid syntax
`
which I first thought to be due to some character encoding issue, given that the line's syntax looks correct. However, if I am not mistaken, it is just that the Python version (2.7) of the particular environment is wrong.

Regarding the license, I was not able to figure out whether it is OSI approved or not. However, this issue seems to have been already handled.

**Manuscript**

The manuscript is overall well-written, concise, and clear. The language is good and sufficient background is provided to put the work in context. However, I would suggest expanding the section on features. In particular, please consider explaining the example algorithm to the reader, as well as providing a description of how complete the software is at the moment. For example, do the included models cover a large part of typical use-cases? How were they selected? It would be very helpful if this somehow allowed the user to evaluate if her specific analysis/use-case is supported or not. I also have a number of questions, which I think do not necessitate more than minor changes to the manuscript:

1. Are the any similar competing frameworks?
2. It is claimed that the union of intersections framework is naturally scalable. However, it is not readily apparent why this is so. Please clarify this point.
3. It would help the reader if it was specified whether the vectors are row or column vectors. This information needs to be now inferred from the equations.
4. What happens if the true beta in equations (1) and (2) is not sufficiently sparse? How robust is the method to slight violations of its assumptions?
5. In the applications section, it would be nice if something was stated about the obtained results. Did PyUoI give better results compared to other approaches when it was used? Also, was there any particular reason to apply the method to neuroscience and genomics data?
6. It is stated that PyUoI is agnostic to the optimization method (solver). Why is this so?
7. I would also suggest adding citations to Pedregosa et al. (2011) and Dalcin et al. (2005). Please see the pages https://scikit-learn.org/stable/about.html#citing-scikit-learn and https://mpi4py.readthedocs.io/en/stable/citing.html for the key references.
