### 2019-11-22 d969e52
updates for install
