+++ requirements.txt
@@ -1,3 +1,3 @@
-numpy
-h5py
-scikit-learn
+numpy>=1.14
+h5py>=2.8
+scikit-learn>=0.20
