### 2019-11-22 6b247c0
updates to paper for response to review
