+++ paper/paper.bib
@@ -71,3 +71,26 @@
   journal={arXiv preprint arXiv:1901.04592},
   year={2019}
 }
+
+@article{dalcin2005,
+  title={MPI for Python},
+  author={Dalc{\'\i}n, Lisandro and Paz, Rodrigo and Storti, Mario},
+  journal={Journal of Parallel and Distributed Computing},
+  volume={65},
+  number={9},
+  pages={1108--1115},
+  year={2005},
+  publisher={Elsevier}
+}
+
+@article{scikit-learn,
+ title={Scikit-learn: Machine Learning in {P}ython},
+ author={Pedregosa, F. and Varoquaux, G. and Gramfort, A. and Michel, V.
+         and Thirion, B. and Grisel, O. and Blondel, M. and Prettenhofer, P.
+         and Weiss, R. and Dubourg, V. and Vanderplas, J. and Passos, A. and
+         Cournapeau, D. and Brucher, M. and Perrot, M. and Duchesnay, E.},
+ journal={Journal of Machine Learning Research},
+ volume={12},
+ pages={2825--2830},
+ year={2011}
+}
