+++ paper/paper.md
@@ -72,9 +72,13 @@ loss in predictive accuracy, and reduced bias relative to benchmark approaches.
 algorithms, encompassing regression, classification, and dimensionality
 reduction. In order to better facilitate its usage, `PyUoI`'s API is structured
 similarly to the `scikit-learn` package, which is a commonly used Python machine
-learning library [@sklearn_api]. Additionally, because the UoI framework is
-naturally scalable, `PyUoI` is equipped with `mpi4py` functionality to
-parallelize model fitting on large datasets.
+learning library [@scikit-learn; @sklearn_api].
+
+The UoI framework operates by fitting many models across resamples of the
+dataset and across a set of regularization parameters. Thus, since these fits
+can be performed in parallel, the UoI framework is naturally scalable. `PyUoI`
+is equipped with `mpi4py` functionality to parallelize model fitting on large
+datasets [@dalcin2005].
 
 # Background
 
@@ -86,27 +90,27 @@ other generalized linear models (logistic and poisson). We refer the user to
 existing literature on the UoI variants of column subset selection and
 non-negative matrix factorization [@bouchard2017; @ubaru2017].
 
-Linear regression consists of estimating parameters $\beta \in \mathbb{R}^p$
-that map a $p$-dimensional vector of features $x \in \mathbb{R}^p$ to the
-observation variable $y\in \mathbb{R}$, when the $N$ samples are corrupted by
-i.i.d Gaussian noise:
+Linear regression consists of estimating parameters $\boldsymbol{\beta} \in
+\mathbb{R}^{p\times 1}$ that map a $p$-dimensional vector of features
+$\mathbf{x} \in \mathbb{R}^{p \times 1}$ to the observation variable $y\in
+\mathbb{R}$, when the $N$ samples are corrupted by i.i.d Gaussian noise:
 
 \begin{equation}
-y = \beta^T x + \epsilon
+y = \boldsymbol{\beta}^T \mathbf{x} + \epsilon
 \end{equation}
 
 where $\epsilon \sim \mathcal{N}(0, \sigma^2)$ for each sample. When the true
-$\beta$ is thought to be sparse (i.e., some subset of the $\beta$ are exactly
-zero), an estimate of $\beta$ can be found by solving a constrained optimization
-problem of the form
+$\boldsymbol{\beta}$ is thought to be sparse (i.e., some subset of the $\beta$
+are exactly zero), an estimate of $\boldsymbol{\beta}$ can be found by solving a
+constrained optimization problem of the form
 
 \begin{equation}
-\hat{\beta} = \underset{\beta\in \mathbb{R}^p}{\text{argmin}}
-                \frac{1}{N}\sum_{i=1}^N(y_i - \beta \cdot x_i)^2
-                + \lambda |\beta|_1
+\hat{\boldsymbol{\beta}} = \underset{\boldsymbol{\beta}\in \mathbb{R}^p}{\text{argmin}}
+                \frac{1}{N}\sum_{i=1}^N(y_i - \boldsymbol{\beta}^T \mathbf{x}_i)^2
+                + \lambda |\boldsymbol{\beta}|_1
 \end{equation}
 
-where $|\beta|_1$ is the $\ell_1$-norm of the parameters. The $\ell_1$-norm is a
+where $|\boldsymbol{\beta}|_1$ is the $\ell_1$-norm of the parameters. The $\ell_1$-norm is a
 convenient penalty because it will tend to force parameters to be set exactly
 equal to zero, performing feature selection [@tibshirani1994]. Typically,
 $\lambda$, the degree to which feature sparsity is enforced, is unknown and must
[…]
