### 2019-11-22 9ee3107
add a couple citations
