+++ paper/paper.bib
@@ -94,3 +94,23 @@
  pages={2825--2830},
  year={2011}
 }
+
+@incollection{akaike1998,
+  title={Information theory and an extension of the maximum likelihood principle},
+  author={Akaike, Hirotogu},
+  booktitle={Selected papers of hirotugu akaike},
+  pages={199--213},
+  year={1998},
+  publisher={Springer}
+}
+
+@article{schwarz1978,
+  title={Estimating the dimension of a model},
+  author={Schwarz},
+  journal={The annals of statistics},
+  volume={6},
+  number={2},
+  pages={461--464},
+  year={1978},
+  publisher={Institute of Mathematical Statistics}
+}
