### 2019-11-26 #1799 @pssachdeva
Hi @puolival,

Thanks for your comments.

**Checklist**

> Please make it visible on the requirement lists and the installation instructions page which version numbers of each used package are supported, and which Python version the user is supposed to have.

As you stated, the Python version may be the source of the errors you encountered. We have opted to not support Python 2.7 since it is losing support in 2020. However, we do not explicitly state this in either the documentation or the README. We have made it explicit in the documentation and README that PyUoI is tested and supported for Python 3.5+.

Additionally, we have clarified the minimum version numbers for the required packages in PyUoI, both in the `requirements.txt` and the documentation.

> Regarding the license, I was not able to figure out whether it is OSI approved or not. However, this issue seems to have been already handled.

We used the BSD-3-Clause-LBNL license, which is OSI approved. See [here](https://opensource.org/BSD-3-Clause-LBNL) for the description on the OSI webpage.

**Manuscript**
> In particular, please consider explaining the example algorithm to the reader

Could you elaborate on this? Our manuscript includes the pseudocode for a detailed description of the algorithm, and the Background section provides a more general description of the framework (with key steps highlighted in the pseudocode for clarity).

> as well as providing a description of how complete the software is at the moment. For example, do the included models cover a large part of typical use-cases? How were they selected? It would be very helpful if this somehow allowed the user to evaluate if her specific analysis/use-case is supported or not.

We have added a paragraph in the Features section clarifying our choice of generalized linear models. We chose these because they are the most commonly used models in a variety of disciplines, particularly neuroscience and bioinformatics which are our active fields of research.

> 1. Are the any similar competing frameworks?

There are competing approaches to obtaining sparse parameter estimates to specific types of models, such as the linear regression model or poisson regression model (e.g. coordinate descent in glmnet). UoI was compared to a battery of other approaches for linear regression in the cited NeurIPS paper.

> 2. It is claimed that the union of intersections framework is naturally scalable. However, it is not readily apparent why this is so. Please clarify this point.

Thank you for pointing this out. We have added the following text in the Summary to better clarify the scalability:

>> The UoI framework operates by fitting many models across resamples of the
dataset and across a set of regularization parameters. Thus, since these fits
can be performed in parallel, the UoI framework is naturally scalable. `PyUoI`
is equipped with `mpi4py` functionality to parallelize model fitting on large
datasets [@dalcin2005].

> 3. It would help the reader if it was specified whether the vectors are row or column vectors. This information needs to be now inferred from the equations.

We have clarified the shapes of the vectors in the Background section to emphasize that they are column vectors. Additionally, we have bolded all vectors to add further clarity.

> 4. What happens if the true beta in equations (1) and (2) is not sufficiently sparse? How robust is the method to slight violations of its assumptions?

We note that the formulation of the Lasso problem makes no constraints on the exact level of sparsity. Indeed, determining the optimal lambda value is often part of the model estimation procedure. Thus, if the true parameters are more dense, the UoI procedure will obtain more dense parameter estimates.

The original Union of Intersections paper examined its performance across a variety of underlying distributions of model parameters, degrees of sparsity, and noise levels, finding superior performance compared to a battery of other algorithms. See the appendix of the paper for more details.

> 5. In the applications section, it would be nice if something was stated about the obtained results. Did PyUoI give better results compared to other approaches when it was used?

The Union of Intersections framework is capable of obtaining sparse and predictive fits for parametric models. We address the capabilities of UoI in the second paragraph of the Summary, where we state that:

>>Models inferred through the UoI framework are characterized by their usage of fewer parameters with little or no loss in predictive accuracy, and reduced bias relative to benchmark approaches.

> 5. Also, was there any particular reason to apply the method to neuroscience and genomics data?

We and our collaborators conduct research in these fields. Neuroscience and genomics datasets are often high-dimensional with structure that is well-suited to the enforcement of a sparsity prior.

> 6. It is stated that PyUoI is agnostic to the optimization method (solver). Why is this so?

This language may have been too strong. The _Union of Intersections_ framework is agnostic to the solver. For example, the UoI Lasso approach requires obtaining (many) solutions to Lasso optimization problems (on resampled datasets), which could be obtained via coordinate descent, gradient descent, orthant-wise LBFGS, etc. The type of solver is not really important insofar as it provides good estimates.

With this in mind, we have structured PyUoI to be extendable to other solvers. For example, UoI Lasso can use a coordinate descent solver (obtained from `scikit-learn`) or the pycasso solver (cited). The user could conceivably extend the UoI framework to use their own choice solver. Thus, _PyUoI_ is not agnostic to the solver without the additional work of extending the software. We have added text in the Features section to clarify this.

> 7. I would also suggest adding citations to Pedregosa et al. (2011) and Dalcin et al. (2005). Please see the pages https://scikit-learn.org/stable/about.html#citing-scikit-learn and https://mpi4py.readthedocs.io/en/stable/citing.html for the key references.

Thank you for bringing this up. We have added these references to the paper.
