### 2019-11-28 #1799 @yzhao062
I just read the manuscript and it looks good to me. The author also addresses my previous concern in the post above. As a result, I am happy to recommend the paper for acceptance at JOSS :)
@pssachdeva @terrytangyuan
