### 2019-12-04 #1799 @terrytangyuan
@pssachdeva Could you fix the missing DOIs mentioned above?
