### 2019-12-04 14d0d83
update version, add py3.8 and remove py3.5
added conda badge

remove 3.8 and paper updates
