### 2019-12-04 0ce008b
Update pyuoi/linear_model/poisson.py
Co-Authored-By: Jesse Livezey <jesse.livezey@gmail.com>
