+++ pyuoi/linear_model/poisson.py
@@ -639,7 +639,7 @@ class UoI_Poisson(AbstractUoIGeneralizedLinearRegressor, Poisson):
             if np.any(y):
                 self.intercept_ = np.log(np.mean(y, keepdims=True) / np.mean(mu))
             else:
-                self.intercept_ = -np.inf * np.ones(1)
+                self.intercept_ = np.array(-np.inf)
         else:
             self.intercept_ = np.zeros(1)
 
