### 2019-12-04 95c20b7
flake8 changes
