### 2019-12-05 #1799 @danielskatz
JOSS wants the DOI associated with the specific version that is being discussed in this paper, not the “concept” or “all versions” DOI
