### 2019-12-05 #1799 @pssachdeva
@terrytangyuan @danielskatz OK, here's the DOI for the version discussed in the paper: 10.5281/zenodo.3563147
