### 2019-12-05 #1799 @terrytangyuan
@danielskatz The paper looks good to me now. Could you take it from here?
