### 2019-12-05 #1799 @danielskatz
Yes, but in the future, please mention @openjournals/jose-eics in such requests
