+++ paper/paper.bib
@@ -1,15 +1,15 @@
 @incollection{bouchard2017,
-    title = {Union of Intersections (UoI) for Interpretable Data Driven Discovery and Prediction},
+    title = {Union of {I}ntersections ({UoI}) for Interpretable Data Driven Discovery and Prediction},
     author = {Bouchard, Kristofer and Bujan, Alejandro and Roosta-Khorasani, Farbod and Ubaru, Shashanka and Prabhat, Mr. and Snijders, Antoine and Mao, Jian-Hua and Chang, Edward and Mahoney, Michael W and Bhattacharya, Sharmodeep},
-    booktitle = {Advances in Neural Information Processing Systems 30},
+    booktitle = {{Advances in Neural Information Processing Systems 30}},
     pages = {1078--1086},
     year = {2017},
 }
 
 @INPROCEEDINGS{ubaru2017,
     author = {S. {Ubaru} and K. {Wu} and K. E. {Bouchard}},
-    booktitle = {2017 16th IEEE International Conference on Machine Learning and Applications (ICMLA)},
-    title = {UoI-NMF Cluster: A Robust Nonnegative Matrix Factorization Algorithm for Improved Parts-Based Decomposition and Reconstruction of Noisy Data},
+    booktitle = {}2017 16th IEEE International Conference on Machine Learning and Applications (ICMLA)}},
+    title = {{UoI-NMF} Cluster: A Robust Nonnegative Matrix Factorization Algorithm for Improved Parts-Based Decomposition and Reconstruction of Noisy Data},
     year = {2017},
     volume = {},
     number = {},
@@ -44,7 +44,7 @@
                Arnaud Joly and Brian Holt and Ga{\"{e}}l Varoquaux},
   title     = {{API} design for machine learning software: experiences from the scikit-learn
                project},
-  booktitle = {ECML PKDD Workshop: Languages for Data Mining and Machine Learning},
+  booktitle = {{ECML PKDD Workshop: Languages for Data Mining and Machine Learning}},
   year      = {2013},
   pages = {108--122},
 }
@@ -52,7 +52,7 @@
 @inproceedings{gong2015,
     author = {Gong, Pinghua and Ye, Jieping},
     title = {A Modified Orthant-Wise Limited Memory Quasi-Newton Method with Convergence Analysis},
-    booktitle = {Proceedings of the 32nd International Conference on International Conference on Machine Learning - Volume 37},
+    booktitle = {{Proceedings of the 32nd International Conference on International Conference on Machine Learning - Volume 37}},
     year = {2015},
     pages = {276--284},
     numpages = {9},
@@ -80,7 +80,7 @@
 }
 
 @article{dalcin2005,
-  title={MPI for Python},
+  title={{MPI} for {P}ython},
   author={Dalc{\'\i}n, Lisandro and Paz, Rodrigo and Storti, Mario},
   journal={Journal of Parallel and Distributed Computing},
   volume={65},
@@ -105,7 +105,7 @@
 @incollection{akaike1998,
   title={Information theory and an extension of the maximum likelihood principle},
   author={Akaike, Hirotogu},
-  booktitle={Selected papers of hirotugu akaike},
+  booktitle={{Selected papers of Hirotugu Akaike}},
   pages={199--213},
   year={1998},
   publisher={Springer},
@@ -115,7 +115,7 @@
 @article{schwarz1978,
   title={Estimating the dimension of a model},
   author={Schwarz},
-  journal={The annals of statistics},
+  journal={The Annals of Statistics},
   volume={6},
   number={2},
   pages={461--464},
