### 2019-12-05 #1799 @danielskatz
@pssachdeva - please merge https://github.com/BouchardLab/pyuoi/pull/193 and check to make sure all the cases are correct in the references.
