### 2019-12-05 8f6bef8
fixed error in paper edits
