+++ paper/paper.bib
@@ -8,7 +8,7 @@
 
 @INPROCEEDINGS{ubaru2017,
     author = {S. {Ubaru} and K. {Wu} and K. E. {Bouchard}},
-    booktitle = {}2017 16th IEEE International Conference on Machine Learning and Applications (ICMLA)}},
+    booktitle = {{2017 16th IEEE International Conference on Machine Learning and Applications (ICMLA)}},
     title = {{UoI-NMF} Cluster: A Robust Nonnegative Matrix Factorization Algorithm for Improved Parts-Based Decomposition and Reconstruction of Noisy Data},
     year = {2017},
     volume = {},
