### 2019-12-06 #1799 @arfon
> 👋 @arfon - please see the XML, ref 7, for the same issue we saw before with the URL
Please let me know how to proceed

I've fixed this manually and deposited the metadata with Crossref. Assuming this is OK then I think we're good to mark this as accepted here and close the review.
