### 2019-12-06 #1799 @whedon
:tada::tada::tada: Congratulations on your paper acceptance! :tada::tada::tada:
